server/requirements.txt
# Server-side requirements — used by Docker.
# These mirror the dependencies in pyproject.toml but are kept separate so
# that the Docker build layer cache works efficiently.

openenv-core>=0.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0
websockets>=12.0
